+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "simply-djs",
-  "version": "2.0.2-dev",
+  "version": "2.0.21-dev",
   "description": "A Simple, Easy and Beginner friendly Discord.js Package for everyone. Uses Discord.js v13",
   "main": "simplydjs.js",
   "typings": "index.d.ts",
@@ -73,6 +73,8 @@
     "canvas": "^2.8.0",
     "discord.js": "^13.1.0",
     "mathjs": "^9.4.4",
+    "ms": "^2.1.3",
+    "quickmongo": "^4.0.0",
     "rss-parser": "^3.12.0"
   }
 }
